Create utilities module
There were functions in products.py that will be useful in other modules
but circular imports pop up if these functions don't have their own module.
